serverSide/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
flask-session
flask-cors
redis
python-dotenv